removed api package in ktensorflow-core module, moved everything to module root level
