+++ ktensorflow-core/src/androidInstrumentedTest/kotlin/InterpreterTest.kt
@@ -1,12 +1,12 @@
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import dev.kursor.ktensorflow.api.Interpreter
-import dev.kursor.ktensorflow.api.InterpreterOptions
-import dev.kursor.ktensorflow.api.ModelDesc
-import dev.kursor.ktensorflow.api.Tensor
-import dev.kursor.ktensorflow.api.TensorDataType
-import dev.kursor.ktensorflow.api.TensorShape
-import dev.kursor.ktensorflow.api.typedData
+import dev.kursor.ktensorflow.Interpreter
+import dev.kursor.ktensorflow.InterpreterOptions
+import dev.kursor.ktensorflow.ModelDesc
+import dev.kursor.ktensorflow.Tensor
+import dev.kursor.ktensorflow.TensorDataType
+import dev.kursor.ktensorflow.TensorShape
+import dev.kursor.ktensorflow.typedData
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
