+++ ktensorflow-core/src/commonMain/kotlin/dev/kursor/ktensorflow/Delegate.kt
@@ -1,4 +1,4 @@
-package dev.kursor.ktensorflow.api
+package dev.kursor.ktensorflow
 
 /**
  * Provides hardware acceleration for inference.
