+++ ktensorflow-core/src/commonMain/kotlin/dev/kursor/ktensorflow/Interpreter.kt
@@ -1,4 +1,4 @@
-package dev.kursor.ktensorflow.api
+package dev.kursor.ktensorflow
 
 /**
  * Interpreter to run inference on a model.
