+++ ktensorflow-core/src/commonMain/kotlin/dev/kursor/ktensorflow/ModelDesc.kt
@@ -1,4 +1,4 @@
-package dev.kursor.ktensorflow.api
+package dev.kursor.ktensorflow
 
 /**
  * Model description class
