+++ ktensorflow-core/src/commonMain/kotlin/dev/kursor/ktensorflow/Tensor.kt
@@ -1,4 +1,4 @@
-package dev.kursor.ktensorflow.api
+package dev.kursor.ktensorflow
 
 import dev.kursor.ktensorflow.impl.TensorImpl
 import dev.kursor.ktensorflow.impl.inferTensorDataType
