+++ ktensorflow-core/src/commonMain/kotlin/dev/kursor/ktensorflow/TensorDataType.kt
@@ -1,4 +1,4 @@
-package dev.kursor.ktensorflow.api
+package dev.kursor.ktensorflow
 
 /**
  * Tensor data types.
