+++ ktensorflow-core/src/iosMain/kotlin/dev/kursor/ktensorflow/TensorFlowException.kt
@@ -1,4 +1,4 @@
-package dev.kursor.ktensorflow.api
+package dev.kursor.ktensorflow
 
 /**
  * Exception thrown by TensorFlow as NSError.
