+++ ktensorflow-gpu/src/iosMain/kotlin/dev/kursor/ktensorflow/gpu/GpuDelegate.kt
@@ -1,7 +1,6 @@
 package dev.kursor.ktensorflow.gpu
 
-import dev.kursor.ktensorflow.api.Delegate
-import dev.kursor.ktensorflow.gpu.IosGpuDelegate
+import dev.kursor.ktensorflow.Delegate
 
 actual fun GpuDelegate(options: GpuDelegateOptions): Delegate {
     return IosGpuDelegate(options.tflOptions)
