+++ ktensorflow-moko/src/iosMain/kotlin/dev/kursor/ktensorflow/moko/MokoModelDesc.kt
@@ -2,7 +2,7 @@ package dev.kursor.ktensorflow.moko
 
 import dev.icerock.moko.resources.AssetResource
 import dev.icerock.moko.resources.FileResource
-import dev.kursor.ktensorflow.api.ModelDesc
+import dev.kursor.ktensorflow.ModelDesc
 
 actual fun ModelDesc.Companion.FileResource(resource: FileResource): ModelDesc {
     return ModelDesc.PathInBundle(resource.path)
