+++ samples/compose-resources/composeApp/src/androidMain/kotlin/dev/kursor/chess/core/model/AndroidModelLoader.kt
@@ -1,7 +1,7 @@
 package dev.kursor.chess.core.model
 
 import android.content.Context
-import dev.kursor.ktensorflow.api.ModelDesc
+import dev.kursor.ktensorflow.ModelDesc
 import ktensorflow.samples.compose_resources.composeapp.generated.resources.Res
 import java.io.File
