+++ samples/compose-resources/composeApp/src/commonMain/kotlin/dev/kursor/chess/core/model/ModelLoader.kt
@@ -1,6 +1,6 @@
 package dev.kursor.chess.core.model
 
-import dev.kursor.ktensorflow.api.ModelDesc
+import dev.kursor.ktensorflow.ModelDesc
 
 interface ModelLoader {
     suspend fun loadModel(): ModelDesc
