+++ samples/moko-resources/composeApp/src/commonMain/kotlin/dev/kursor/chess/features/game/data/ChessAiMoveRepositoryImpl.kt
@@ -9,15 +9,15 @@ import dev.kursor.chess.engine.model.Color
 import dev.kursor.chess.engine.model.PieceType
 import dev.kursor.chess.engine.model.Square
 import dev.kursor.chess.features.game.domain.ChessAiMoveRepository
-import dev.kursor.ktensorflow.api.Interpreter
-import dev.kursor.ktensorflow.api.InterpreterOptions
-import dev.kursor.ktensorflow.api.ModelDesc
-import dev.kursor.ktensorflow.api.Tensor
-import dev.kursor.ktensorflow.api.TensorDataType
-import dev.kursor.ktensorflow.api.TensorShape
+import dev.kursor.ktensorflow.Interpreter
+import dev.kursor.ktensorflow.InterpreterOptions
+import dev.kursor.ktensorflow.ModelDesc
+import dev.kursor.ktensorflow.Tensor
+import dev.kursor.ktensorflow.TensorDataType
+import dev.kursor.ktensorflow.TensorShape
 import dev.kursor.ktensorflow.gpu.GpuDelegate
-import dev.kursor.ktensorflow.api.typedData
 import dev.kursor.ktensorflow.moko.FileResource
+import dev.kursor.ktensorflow.typedData
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
